Added Makezip
moved from fmt to std

moved each file into a folder
