--- src/underDevelopment/mainzip.cpp
@@ -1,1 +1,0 @@
-// this is a project to make a zip based on make with the command makezip
